changed payment styling removed from nav
