+++ components/Payment/Payment.module.css
@@ -20,3 +20,8 @@
     justify-content: space-evenly;
     align-items: center;
 }
+
+.discount {
+    color: lime;
+    float: right;
+}
