+++ components/Premium/BuyPremium/BuyPremium.module.css
@@ -1,0 +1,5 @@
+.label {
+    width: 300px;
+    float: left;
+    margin: 0;
+}
